chore/Refactor aad user fetch/update process
- Added new adaterps, tasks modules in azure model.
- Remove unnecessary code to prevent circular imports.
- Clean up azure classes to increase maintanability and separation of concerns.
