+++ django/user/tasks.py
@@ -2,7 +2,6 @@ from celery.utils.log import get_task_logger
 
 from core.utils import send_mail_wrapper
 from scheduler.celery import app
-from .adapters import MyAzureAccountAdapter
 
 logger = get_task_logger(__name__)
 
@@ -50,15 +49,3 @@ def send_user_request_to_admins(profile_id):
                           to=admin.user.email,
                           language=admin.language,
                           context=context)
-
-
-@app.task(name="fetch_users_from_aad_and_update_db")
-def fetch_users_from_aad_and_update_db():
-    adapter = MyAzureAccountAdapter()
-    # Fetch the AAD users
-    azure_users = adapter.get_aad_users()
-    # Save the AAD users to the local database and get the updated user profiles
-    updated_user_profiles = adapter.save_aad_users(azure_users)
-
-    logger.info(f"Updated {len(updated_user_profiles)} user profiles")
-    return f'Updated {len(updated_user_profiles)} user profiles. Azure users fetched and updated in the database successfully.'
