+++ django/user/views.py
@@ -1,14 +1,9 @@
 from rest_framework.mixins import CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin
 from rest_framework.viewsets import GenericViewSet
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from rest_framework import status
 
 from core.views import TokenAuthMixin
 from .serializers import UserProfileSerializer, OrganisationSerializer, UserProfileListSerializer
 from .models import UserProfile, Organisation
-from .adapters import MyAzureAccountAdapter
-from user.tasks import fetch_users_from_aad_and_update_db
 
 
 class UserProfileViewSet(TokenAuthMixin, RetrieveModelMixin, UpdateModelMixin, GenericViewSet):
@@ -39,33 +34,3 @@ class OrganisationViewSet(TokenAuthMixin, CreateModelMixin, ListModelMixin, Retr
             return Organisation.objects.filter(name__contains=search_term)
         else:
             return Organisation.objects.all()
-
-
-class GetAADUsers(TokenAuthMixin, APIView):
-    """
-    API View to fetch Azure Active Directory (AAD) users.
-    Requires token authentication.
-    """
-
-    def get(self, request, format=None):
-        # Create an instance of MyAzureAccountAdapter and fetch the AAD users
-        adapter = MyAzureAccountAdapter()
-        azure_users = adapter.get_aad_users()
-
-        # Return the AAD users in the response
-        return Response({'users': azure_users}, status=status.HTTP_200_OK)
-
-
-
-class UpdateAADUsersView(TokenAuthMixin, APIView):
-    """
-    API View to update and save Azure Active Directory (AAD) users in the local database.
-    It fetches the AAD users, saves them, and returns the updated users' profiles.
-    Requires token authentication.
-    """
-
-    def put(self, request, format=None):
-        # Call the Celery task to fetch and update the users
-        fetch_users_from_aad_and_update_db.delay()
-
-        return Response({'message': 'Azure users update started.'}, status=status.HTTP_202_ACCEPTED)
